change `LegislationParser` Document.Body to `Contents`
To make `LegislationParser more useable, it's necessary to separate out
the DOCX file's body into a list of `IBlock` that's easier for the
parser to work with.

This also means indexing into the `Block` Attribute of `BlockWithBreak`
is not longer necessary since `BlockWithBreak` is discarded - the
legislation parsing doesn't require the information. (If it did then we
would want a different method of tracking this information, probably as
an attribute on `IBlock` itself.

rename LegislationParser.Contents to Body